add vault status and gitlab token status

+++ actions/deploy/k8s/scripts/config.sh
@@ -8,6 +8,7 @@ VAULT_TOKEN=$VAULT_TOKEN
 VAULT_SERVER=$VAULT_SERVER
 VAULT_SECRET_PATH=$VAULT_SECRET_PATH
 VAULT_SECRET_COMMON_PATH=$VAULT_SECRET_COMMON_PATH
+VAULT_API=$VAULT_API
 GITLAB_TOKEN=$GITLAB_TOKEN
 GITLAB_SERVER=$GITLAB_SERVER
 GITLAB_PROJECT_NUMBER=$GITLAB_PROJECT_NUMBER
@@ -30,39 +31,54 @@ function configmap() {
     kubectl create configmap app-local-settings --from-file=${LOCAL_SETTINGS_NAME} --namespace=${NAMESPACE}
   fi
 }
-if [ "${ECR_REPOSITORY}" == "python-django "]; then
-    if [ "${LOCAL_SETTINGS}" == "true" ]; then
-      echo "Downloading Latest Settings file..."
-      curl -H "PRIVATE-TOKEN: ${GITLAB_TOKEN}" "${GITLAB_SERVER}/api/v4/projects/${GITLAB_PROJECT_NUMBER}/repository/files/${ENVIRONMENT}%2F${LOCAL_SETTINGS_NAME}/raw?ref=main" -o ${LOCAL_SETTINGS_NAME}
-      echo "Creating settings configmap..."
-      configmap ${LOCAL_SETTINGS_NAME}
-    elif [ "${LOCAL_SETTINGS}" == "false" ]; then
-      echo "Dummy Settings" > settings.txt
-      echo "Creating dummy settings configmap"
-      configmap settings.txt
-    fi
-else  
-  echo "Skipping ConfigMap"
+# Get Gitlab Token Status. If 200, valid token else expired or wrong token
+GITLAB_TOKEN_STATUS=$(curl -s -o /dev/null -w "%{http_code}" --header "PRIVATE-TOKEN: $GITLAB_TOKEN" "https://$GITLAB_SERVER/api/v4/user")
+
+if [ "${GITLAB_TOKEN_STATUS}" == "200" ]; then
+  if [ "${ECR_REPOSITORY}" == "python-django "]; then
+      if [ "${LOCAL_SETTINGS}" == "true" ]; then
+        echo "Downloading Latest Settings file..."
+        curl -H "PRIVATE-TOKEN: ${GITLAB_TOKEN}" "${GITLAB_SERVER}/api/v4/projects/${GITLAB_PROJECT_NUMBER}/repository/files/${ENVIRONMENT}%2F${LOCAL_SETTINGS_NAME}/raw?ref=main" -o ${LOCAL_SETTINGS_NAME}
+        echo "Creating settings configmap..."
+        configmap ${LOCAL_SETTINGS_NAME}
+      elif [ "${LOCAL_SETTINGS}" == "false" ]; then
+        echo "Dummy Settings" > settings.txt
+        echo "Creating dummy settings configmap"
+        configmap settings.txt
+      fi
+  else  
+    echo "Skipping ConfigMap"
+  fi
+else
+  exit 1
 fi
 
-if [ "${ECR_REPOSITORY}" == "python-django" ]; then
-    echo "Extracting variables from Vault..."
-    curl -H "X-Vault-Token: ${VAULT_TOKEN}" "https://${VAULT_SERVER}/v1/${VAULT_SECRET_PATH}" | jq -r .data > .env.json
-    curl -H "X-Vault-Token: ${VAULT_TOKEN}" "https://${VAULT_SERVER}/v1/${VAULT_SECRET_COMMON_PATH}" | jq -r .data > .common.env.json
-    cat .env.json | jq -r 'to_entries[] | "\(.key)=\(.value)"' | base64 > secrets
-    cat .common.env.json | jq -r 'to_entries[] | "\(.key)=\(.value)"' | base64 >> secrets
+# Check for Vault Status. If 200, create secrets else exit
+VAULT_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -H "X-Vault-Token: ${VAULT_TOKEN}" \
+              --insecure https://${VAULT_SERVER}/v1/${VAULT_API})
 
-    if kubectl get secret app-secret --namespace=${NAMESPACE} &> /dev/null; then
-      echo "Creating secrets from variables..."
-      kubectl delete secret app-secret --namespace=${NAMESPACE}
-      kubectl create secret generic app-secret --from-file=secrets --namespace=${NAMESPACE}
-    else
-      echo "Creating secrets from variables..."
-      kubectl create secret generic app-secret --from-file=secrets --namespace=${NAMESPACE}
-    fi
-else
-   echo "The current chart is frontend"
-fi 
+if [ "${VAULT_STATUS}" == "200" ]; then
+  if [ "${ECR_REPOSITORY}" == "python-django" ]; then
+      echo "Extracting variables from Vault..."
+      curl -H "X-Vault-Token: ${VAULT_TOKEN}" "https://${VAULT_SERVER}/v1/${VAULT_SECRET_PATH}" | jq -r .data > .env.json
+      curl -H "X-Vault-Token: ${VAULT_TOKEN}" "https://${VAULT_SERVER}/v1/${VAULT_SECRET_COMMON_PATH}" | jq -r .data > .common.env.json
+      cat .env.json | jq -r 'to_entries[] | "\(.key)=\(.value)"' | base64 > secrets
+      cat .common.env.json | jq -r 'to_entries[] | "\(.key)=\(.value)"' | base64 >> secrets
+
+      if kubectl get secret app-secret --namespace=${NAMESPACE} &> /dev/null; then
+        echo "Creating secrets from variables..."
+        kubectl delete secret app-secret --namespace=${NAMESPACE}
+        kubectl create secret generic app-secret --from-file=secrets --namespace=${NAMESPACE}
+      else
+        echo "Creating secrets from variables..."
+        kubectl create secret generic app-secret --from-file=secrets --namespace=${NAMESPACE}
+      fi
+  else
+     echo "The current chart is frontend"
+  fi
+else  
+  exit 1
+fi
 
 if kubectl get secret kcr-secret --namespace=${NAMESPACE} &> /dev/null; then
   echo "Creating kcr-secrets from variables..."
